Fixed the diminishing science value of transmitted data
Breaking Grounds replaced the implementation of ResearchAndDevelopment.GetReferenceDataValue
with a function that no longer returns a constant value for equally sized data chunks. Instead,
the value of subsequent chunks are less than the values of the previous ones. Usually this
doesn't affect the game at all, because science usually is done in one chunk.

They probably did this for their surface experiments, so that they never cease to produce
scientific value, but will continue to give science output forever.

Fixes #421